+++ @xen-orchestra/lite/src/lang/en.json
@@ -1,5 +1,5 @@
 {
   "confirmCtrlAltDel": "Send Ctrl+Alt+Del to VM?",
   "connect": "Connect",
-  "ctrlAltDel": "Ctrl + Alt + Del",
+  "ctrlAltDel": "Ctrl + Alt + Del"
 }
